Add service management and revenue calculation updates

See merge request vanhaitsu/chillde-api!397

+++ Chillde.Repositories/Models/PackageFeatureModels/PackageFeatureModel.cs
@@ -7,7 +7,7 @@ namespace Chillde.Repositories.Models.PackageFeatureModels
     {
         public string? Name { get; set; }
         public decimal? AdditionalCost { get; set; }
-        public int? AdditionalDay { get; set; }
+        public float? AdditionalDay { get; set; }
         public bool? IsExtra { get; set; }
         public bool? IsChecked { get; set; }
         public int MinQuantity { get; set; }
